requirements.txt
fastapi
uvicorn
rembg
pillow
onnxruntime
python-multipart